+++ projects/ids-enterprise-typings/lib/datagrid/soho-datagrid.d.ts
@@ -1001,7 +1001,7 @@ interface SohoDataGridColumnNumberFormat {
 }
 
 type SohoDataGridColumnNumberFormatStyle = 'decimal' | 'currency' | 'percent' | 'integer' | string;
-type SohoDataGridTriggerSourcePagerType = 'initial' | 'refresh' | 'filtered' | 'sorted' | 'updatecolums' | string;
+type SohoDataGridTriggerSourcePagerType = 'initial' | 'refresh' | 'filtered' | 'sorted' | 'updatecolums' | string | object;
 
 interface SohoGridCellOption {
   /** The underlying data value. */
@@ -1385,7 +1385,7 @@ interface SohoDataGridSettingsChangedEvent {
   rowHeight?: SohoDataGridRowHeight;
   sortOrder?: { columnId: string, ascending?: boolean };
   pagesize?: number;
-  activePage?: string;
+  activePage?: string | number;
   filter?: Array<SohoDataGridFilterCondition>;
 }